#60 Move conversion functions to parsers module
