+++ pandas_dataclasses/__init__.py
@@ -2,6 +2,7 @@
 from . import core
 from .core.frame import *
 from .core.generic import *
+from .core.parsers import *
 from .core.series import *
 from .core.typing import *
