+++ pandas_dataclasses/core/__init__.py
@@ -1,6 +1,7 @@
 # type: ignore
 from . import frame
 from . import generic
-from . import specs
+from . import parsers
 from . import series
+from . import specs
 from . import typing
